requirements-core.txt
numpy>=2.0,<3
Pillow>=10,<12
mpmath>=1.3,<2
tqdm>=4.66,<5
gmpy2
